Bumped version. Added database upgrade for older versions.

+++ src/de/smasi/tickmate/database/DatabaseOpenHelper.java
@@ -3,6 +3,7 @@ package de.smasi.tickmate.database;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.util.Log;
 
 public class DatabaseOpenHelper extends SQLiteOpenHelper {
 	private static DatabaseOpenHelper sharedInstance;
@@ -25,7 +26,7 @@ public class DatabaseOpenHelper extends SQLiteOpenHelper {
     public static final String COLUMN_MULTIPLE_ENTRIES_PER_DAY = "multiple_entries_per_day";
 
     private static final String DATABASE_NAME = "tickmate.db";
-    private static final int DATABASE_VERSION = 9;
+    private static final int DATABASE_VERSION = 10;
     
     public DatabaseOpenHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -62,6 +63,7 @@ public class DatabaseOpenHelper extends SQLiteOpenHelper {
     
 	@Override
 	public void onCreate(SQLiteDatabase db) {
+		Log.d("tickmate", "Creating database");
 		db.execSQL(DATABASE_CREATE_TRACKS);
 		db.execSQL(DATABASE_CREATE_TICKS);
 	}
@@ -69,8 +71,17 @@ public class DatabaseOpenHelper extends SQLiteOpenHelper {
 	@Override
 	public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 		/* TODO: Alter table depending on version number to support older versions */
-	    db.execSQL("DROP TABLE IF EXISTS " + TABLE_TRACKS);
-	    db.execSQL("DROP TABLE IF EXISTS " + TABLE_TICKS);
-	    onCreate(db);
+		Log.d("tickmate", "Upgrading database");
+		if (oldVersion == 9 && newVersion == 10) {
+			Log.d("tickmate", "Adding columns from version 10");
+			db.execSQL("ALTER TABLE " + TABLE_TRACKS + " ADD COLUMN " + COLUMN_MULTIPLE_ENTRIES_PER_DAY + " integer DEFAULT 0;");
+			db.execSQL("ALTER TABLE " + TABLE_TICKS + " ADD COLUMN " + COLUMN_HOUR + " integer;");
+			db.execSQL("ALTER TABLE " + TABLE_TICKS + " ADD COLUMN " + COLUMN_MINUTE + " integer;");
+			db.execSQL("ALTER TABLE " + TABLE_TICKS + " ADD COLUMN " + COLUMN_SECOND + " integer;");
+		} else {
+			db.execSQL("DROP TABLE IF EXISTS " + TABLE_TRACKS);
+		    db.execSQL("DROP TABLE IF EXISTS " + TABLE_TICKS);
+		    onCreate(db);
+		}
 	}
 }
